requirements.txt.txt
opencv-python=4.11.0.86
cvzone=1.1
pyserial=3.5
mediapipe=0.10.21